+++ src/test/java/io/supertokens/test/ThirdPartySignInUpAPITest2_8.java
@@ -18,6 +18,7 @@ package io.supertokens.test;
 
 import com.google.gson.JsonObject;
 import io.supertokens.ProcessState;
+import io.supertokens.emailverification.EmailVerification;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.storageLayer.StorageLayer;
 import org.junit.AfterClass;
@@ -37,7 +38,7 @@ import static org.junit.Assert.*;
  *  - all error states
  * */
 
-public class SignInUpAPITest2_7 {
+public class ThirdPartySignInUpAPITest2_8 {
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -66,10 +67,16 @@ public class SignInUpAPITest2_7 {
         }
 
         JsonObject response = Utils
-                .signInUpRequest_2_7(process, "test@example.com", false, "testThirdPartyId",
+                .signInUpRequest_2_8(process, "test@example.com", "testThirdPartyId",
                         "testThirdPartyUserId");
         checkSignInUpResponse(response, "testThirdPartyId", "testThirdPartyUserId", "test@example.com", true);
 
+        {
+            JsonObject user = response.getAsJsonObject("user");
+            assertFalse(EmailVerification
+                    .isEmailVerified(process.getProcess(), user.get("id").getAsString(),
+                            user.get("email").getAsString()));
+        }
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
@@ -88,12 +95,12 @@ public class SignInUpAPITest2_7 {
         }
 
         JsonObject response_1 = Utils
-                .signInUpRequest_2_7(process, "TeSt@example.com", false, "testThirdPartyId",
+                .signInUpRequest_2_8(process, "TeSt@example.com", "testThirdPartyId",
                         "testThirdPartyUserId");
         checkSignInUpResponse(response_1, "testThirdPartyId", "testThirdPartyUserId", "test@example.com", true);
 
         JsonObject response_2 = Utils
-                .signInUpRequest_2_7(process, "test@example.com", false, "testThirdPartyId",
+                .signInUpRequest_2_8(process, "test@example.com", "testThirdPartyId",
                         "testThirdPartyUserId");
         checkSignInUpResponse(response_2, "testThirdPartyId", "testThirdPartyUserId", "test@example.com", false);
 
@@ -125,7 +132,7 @@ public class SignInUpAPITest2_7 {
                         .sendJsonPOSTRequest(process.getProcess(), "",
                                 "http://localhost:3567/recipe/signinup", null, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_7ForTests(), "thirdparty");
+                                null, Utils.getCdiVersion2_8ForTests(), "thirdparty");
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -148,7 +155,7 @@ public class SignInUpAPITest2_7 {
                         .sendJsonPOSTRequest(process.getProcess(), "",
                                 "http://localhost:3567/recipe/signinup", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_7ForTests(), "thirdparty");
+                                null, Utils.getCdiVersion2_8ForTests(), "thirdparty");
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -169,7 +176,7 @@ public class SignInUpAPITest2_7 {
                         .sendJsonPOSTRequest(process.getProcess(), "",
                                 "http://localhost:3567/recipe/signinup", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_7ForTests(), "thirdparty");
+                                null, Utils.getCdiVersion2_8ForTests(), "thirdparty");
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -211,32 +218,12 @@ public class SignInUpAPITest2_7 {
                         .sendJsonPOSTRequest(process.getProcess(), "",
                                 "http://localhost:3567/recipe/signinup", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_7ForTests(), "thirdparty");
+                                null, Utils.getCdiVersion2_8ForTests(), "thirdparty");
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
                         e.getMessage()
                                 .equals("Http error. Status Code: 400. Message: Field name 'id' is " +
-                                        "invalid " +
-                                        "in " +
-                                        "JSON input"));
-            }
-        }
-        {
-            JsonObject emailObject = new JsonObject();
-            emailObject.addProperty("id", "test@example.com");
-            requestBody.add("email", emailObject);
-            try {
-                io.supertokens.test.httpRequest.HttpRequest
-                        .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/signinup", requestBody, 1000,
-                                1000,
-                                null, Utils.getCdiVersion2_7ForTests(), "thirdparty");
-                throw new Exception("Should not come here");
-            } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
-                assertTrue(e.statusCode == 400 &&
-                        e.getMessage()
-                                .equals("Http error. Status Code: 400. Message: Field name 'isVerified' is " +
                                         "invalid " +
                                         "in " +
                                         "JSON input"));